refactor: Move rendering methods to MarkdownRenderUtil
